+++ R/fit_g_mc_variableJ.R
@@ -32,7 +32,7 @@
 #' #hence the complicated argument in read.csv()
 #' #This dataset is a CO2 by light response curve for a single sunflower
 #' data <- read.csv(system.file("extdata", "A_Ci_Q_data_1.csv",
-#'                              package = "plantecophystools"))
+#'                              package = "photosynthesis"))
 #'
 #' #Note: there will be issues here if the alpha value used
 #' #for calculating ETR is off, if gamma_star is incorrect,