+++ R/print_graphs.R
@@ -20,13 +20,13 @@
 #' @importFrom graphics plot
 #' @export
 #'
-#' @examples \dontrun{
+#' @examples \donttest{
 #' #Read in your data
 #' #Note that this data is coming from data supplied by the package
 #' #hence the complicated argument in read.csv()
 #' #This dataset is a CO2 by light response curve for a single sunflower
 #' data <- read.csv(system.file("extdata", "A_Ci_Q_data_1.csv",
-#'                              package = "plantecophystools"))
+#'                              package = "photosynthesis"))
 #'
 #' #Fit many AQ curves
 #' #Set your grouping variable
@@ -74,10 +74,11 @@
 #'                             list_element = 3)
 #'
 #' #Print graphs to pdf
-#' print_graphs(data = fits_graphs,
-#'              output_type = "pdf",
-#'              path = tempdir(),
-#'              pdf_filename = "mygraphs.pdf")
+#' #Uncomment to run
+#' #print_graphs(data = fits_graphs,
+#' #            output_type = "pdf",
+#' #            path = tempdir(),
+#' #            pdf_filename = "mygraphs.pdf")
 #' }
 print_graphs <- function(data,
                          path,