+++ R/t_functions.R
@@ -1,6 +1,6 @@
 #' Temperature response functions
 #'
-#' @param Tleaf Leaf temperature in Celsius
+#' @param T_leaf Leaf temperature in K
 #'
 #' @param a Constant to minimize residuals (Heskel et al. 2016)
 #' @param b Linear coefficient to minmize residuals (Heskel et al. 2016)
@@ -15,7 +15,7 @@
 #' @param dS Entropy parameter in J mol-1 (Medlyn et al. 2002)
 #' @param Ea Activation energy in J mol-1 (Medlyn et al. 2002)
 #' @param Hd Deactivation energy in J mol-1 (Medlyn et al. 2002)
-#' @param Topt Optimum temperature of the process in Celsius (Medlyn et al.
+#' @param Topt Optimum temperature of the process in K (Medlyn et al.
 #' 2002)
 #'
 #' @param dH Change in enthalpy of the reaction at 25 C in J mol-1 (Hobbs et
@@ -72,9 +72,9 @@
 #'
 #' @rdname t_functions
 #' @export
-t_response_arrhenius <- function(Tleaf, Ea) {
-  exp(Ea * ((Tleaf + 273.15) - 298.15) /
-        (298.15 * 8.314 * (Tleaf + 273.15)))
+t_response_arrhenius <- function(T_leaf, Ea) {
+  exp(Ea * ((T_leaf) - 298.15) /
+        (298.15 * 8.314 * (T_leaf)))
 }
 
 #' @rdname t_functions
@@ -85,20 +85,20 @@ t_response_arrhenius_kruse <- function(dEa, Ea_ref, Par_ref, T2) {
 
 #' @rdname t_functions
 #' @export
-t_response_arrhenius_medlyn <- function(Tleaf, Ea, Hd, dS) {
-  exp(Ea * ((Tleaf + 273.15) - 298.15) /
-        (298.15 * 8.314 * (Tleaf + 273.15))) *
+t_response_arrhenius_medlyn <- function(T_leaf, Ea, Hd, dS) {
+  exp(Ea * ((T_leaf) - 298.15) /
+        (298.15 * 8.314 * (T_leaf))) *
     (1 + exp((298.15 * dS - Hd) / (298.15 * 8.314))) /
-    (1 + exp(((Tleaf + 273.15) * dS - Hd) / ((Tleaf + 273.15) * 8.314)))
+    (1 + exp(((T_leaf) * dS - Hd) / ((T_leaf) * 8.314)))
 }
 
 #' @rdname t_functions
 #' @export
-t_response_arrhenius_topt <- function(Tleaf, Ea, Hd, Topt) {
-  Hd * exp(Ea * ((Tleaf + 273.15) - (Topt + 273.15)) /
-             ((Tleaf + 273.15) * (Topt + 273.15) * 8.314)) /
-    (Hd - Ea * (1 - exp(Hd * ((Tleaf + 273.15) - (Topt + 273.15)) /
-                          ((Tleaf + 273.15) * (Topt + 273.15) * 8.314))))
+t_response_arrhenius_topt <- function(T_leaf, Ea, Hd, Topt) {
+  Hd * exp(Ea * ((T_leaf) - (Topt)) /
+             ((T_leaf) * (Topt) * 8.314)) /
+    (Hd - Ea * (1 - exp(Hd * ((T_leaf) - (Topt)) /
+                          ((T_leaf) * (Topt) * 8.314))))
 }
 
 #' @rdname t_functions
@@ -117,8 +117,8 @@ t_response_calc_topt <- function(Hd, dS, Ea) {
 
 #' @rdname t_functions
 #' @export
-t_response_heskel <- function(Tleaf, a, b, c) {
-  a + b * (Tleaf) + c * (Tleaf) ^ 2
+t_response_heskel <- function(T_leaf, a, b, c) {
+  a + b * (T_leaf - 273.15) + c * (T_leaf - 273.15) ^ 2
 }
 
 #' @rdname t_functions
@@ -126,9 +126,9 @@ t_response_heskel <- function(Tleaf, a, b, c) {
 t_response_mmrt <- function(dCp,
                             dG,
                             dH,
-                            Tleaf) {
+                            T_leaf) {
   (log(1.380649e-23 * (298.15) / 6.62607e-34)) -
     dG / (8.314 * 298.15) +
-    (1 / 298.15 + dH / (8.314 * 298.15 ^ 2)) * ((Tleaf + 273.15) - 298.15) +
-    (dCp / (2 * 8.314 * 298.15 ^ 2)) * ((Tleaf + 273.15) - 298.15) ^ 2
+    (1 / 298.15 + dH / (8.314 * 298.15 ^ 2)) * ((T_leaf) - 298.15) +
+    (dCp / (2 * 8.314 * 298.15 ^ 2)) * ((T_leaf) - 298.15) ^ 2
 }