+++ tests/testthat/test-fit_aci_response.R
@@ -3,7 +3,7 @@ library(photosynthesis)
 context("Fitting ACi curves")
 
 df <- data.frame(A_net = c(15, 10, 6.5, 2, -0.2, 14.5, 19, 21, 22.5, 23),
-                 Tleaf = c(25, 25, 25, 25, 25, 25, 25, 25, 25, 25),
+                 T_leaf = rep(298, 10),
                  C_i = c(320, 240, 170, 90, 50, 340, 510, 710, 1070, 1450),
                  PPFD = c(rep(1500, 10)))